Horizontal scrolling and depth working!
TODO:
	- change metrics to int. Only derivatives need to be float.
	- vertical scrolling.
	- some kind of mask for better depth rendering.

+++ app/src/main/java/com/learn/flavio_mauricio/beatemupgame/logic/Floor.java
@@ -18,8 +18,8 @@ public class Floor extends GameObject {
         return floorLimit;
     }
 
-    public float getSizeX() {
-        return sizeX;
+    public int getSizeX() {
+        return (int) sizeX;
     }
 
 }
